Determine the Pooling Strategy for an Imputation Method

Maps an rbmi method object to the pooling strategy that rbmi::pool() will apply to the analysis results:

rbmi::method_bayes() / rbmi::method_approxbayes() -> "rubin"

rbmi::method_condmean() with type = "bootstrap" -> "bootstrap"

rbmi::method_condmean() with type = "jackknife" -> "jackknife"

rbmi::method_bmlmi() -> "bmlmi"

This is the single source of truth used internally by analyse_mi_data(); it is exported so the mapping is transparent and usable in user code (see issue #50).

Usage

get_pooling(method)

Arguments

method

A method object created by rbmi::method_bayes(), rbmi::method_approxbayes(), rbmi::method_condmean(), or rbmi::method_bmlmi().

Value

A length-one character vector: "rubin", "bootstrap", "jackknife", or "bmlmi".

See also

Examples

get_pooling(rbmi::method_bayes(n_samples = 10))
#> [1] "rubin"
get_pooling(rbmi::method_condmean(type = "jackknife"))
#> [1] "jackknife"