rbmiUtils: Utility Functions to Support and Extend the 'rbmi' Package

Provides utility functions that extend the capabilities of the reference-based multiple imputation package 'rbmi'. It supports clinical trial analysis workflows with functions for managing imputed datasets, applying analysis methods across imputations, and tidying results for reporting.

See also

Author

Maintainer: Mark Baillie bailliem@gmail.com (ORCID) [copyright holder]

Authors:

Tobias Mütze tobias.muetze@novartis.com (ORCID)

Jack Talboys jack.talboys@novartis.com

Other contributors:

Lukas A. Widmer (ORCID) [contributor]